+++ .storybook/config.js
@@ -1,9 +1,10 @@
 import { configure } from '@storybook/react';
 import '@storybook/addon-actions/register';
 
+// automatically import all files ending in *.stories.js
+const req = require.context('../stories', true, /.stories.js$/);
 function loadStories() {
-  require('../stories/index.js');
-  // You can require as many stories as you need.
+  req.keys().forEach(filename => req(filename));
 }
 
 configure(loadStories, module);
